pyproject.toml
[project]
name = "localragchat"
version = "0.1.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "bitsandbytes>=0.48.1",
    "fastapi>=0.115.0",
    "huggingface-hub>=0.35.3",
    "protobuf>=6.32.1",
    "psycopg[binary]>=3.2.10",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "sentencepiece>=0.2.1",
    "tokenizers>=0.22.1",
    "torch==2.7.1",
    "torchaudio==2.7.1",
    "torchvision==0.22.1",
    "transformers>=4.57.0",
    "uvicorn>=0.32.0",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchaudio = { index = "pytorch" }
torchvision = { index = "pytorch" }
